refactors reduce method

+++ flow.go
@@ -9,20 +9,20 @@ type (
 	}
 
 	flowControl struct {
-		concurrency uint8
+		concurrency uint32
 	}
 
 	FlowOption func(fc *flowControl)
 )
 
-func newFlow[K constraints.Ordered, V any](concurrency uint8) *flow[K, V] {
+func newFlow[K constraints.Ordered, V any](concurrency uint32) *flow[K, V] {
 	return &flow[K, V]{
 		ch:   make(chan OrderedPair[K, V]),
 		stop: make(chan struct{}, concurrency),
 	}
 }
 
-func Concurrency(n uint8) FlowOption {
+func Concurrency(n uint32) FlowOption {
 	return func(fc *flowControl) {
 		fc.concurrency = n
 	}
